+++ src/main/java/com/xiaoleilu/hutool/setting/AbsSetting.java
@@ -5,18 +5,18 @@ import java.nio.charset.Charset;
 import java.util.Map;
 
 import com.xiaoleilu.hutool.getter.OptNullBasicTypeFromObjectGetter;
+import com.xiaoleilu.hutool.util.CharsetUtil;
 
 /**
- * 设定抽象类
+ * Setting抽象类
  * @author Looly
  *
  */
 public abstract class AbsSetting extends OptNullBasicTypeFromObjectGetter<String> implements Map<Object, Object>{
 	
-	/** 本设置对象的字符集 */
-	protected Charset charset;
-	/** 是否使用变量 */
-	protected boolean isUseVariable;
+	/** 默认字符集 */
+	public final static Charset DEFAULT_CHARSET = CharsetUtil.CHARSET_UTF_8;
+	
 	/** 设定文件的URL */
 	protected URL settingUrl;