+++ src/main/java/com/xiaoleilu/hutool/setting/dialect/BasicSetting.java
@@ -1,17 +1,10 @@
 package com.xiaoleilu.hutool.setting.dialect;
 
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.Writer;
 import java.net.URL;
 import java.nio.charset.Charset;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.Map;
 import java.util.Properties;
@@ -19,16 +12,13 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
 import com.xiaoleilu.hutool.convert.Convert;
-import com.xiaoleilu.hutool.io.FileUtil;
-import com.xiaoleilu.hutool.io.IoUtil;
 import com.xiaoleilu.hutool.log.Log;
 import com.xiaoleilu.hutool.log.StaticLog;
 import com.xiaoleilu.hutool.setting.AbsSetting;
 import com.xiaoleilu.hutool.setting.Setting;
+import com.xiaoleilu.hutool.setting.SettingLoader;
 import com.xiaoleilu.hutool.util.BeanUtil;
 import com.xiaoleilu.hutool.util.BeanUtil.ValueProvider;
-import com.xiaoleilu.hutool.util.CharsetUtil;
-import com.xiaoleilu.hutool.util.ReUtil;
 import com.xiaoleilu.hutool.util.StrUtil;
 import com.xiaoleilu.hutool.util.URLUtil;
 
@@ -45,40 +35,19 @@ import com.xiaoleilu.hutool.util.URLUtil;
 public class BasicSetting extends AbsSetting{
 	private final static Log log = StaticLog.get();
 	
+	final private LinkedList<String> groups = new LinkedList<String>();
 	final Map<Object, Object> map = new ConcurrentHashMap<>();
 	
-	/** 默认字符集 */
-	public final static Charset DEFAULT_CHARSET = CharsetUtil.CHARSET_UTF_8;
-	/** 数组类型值默认分隔符 */
-	public final static String DEFAULT_DELIMITER= ",";
+	/** 本设置对象的字符集 */
+	protected Charset charset;
+	/** 是否使用变量 */
+	protected boolean isUseVariable;
 
-	/** 注释符号（当有此符号在行首，表示此行为注释） */
-	private final static String COMMENT_FLAG_PRE = "#";
-	/** 赋值分隔符（用于分隔键值对） */
-	private final static String ASSIGN_FLAG = "=";
-	/** 分组行识别的环绕标记 */
-	private final static char[] GROUP_SURROUND = { '[', ']' };
-	/** 变量名称的正则 */
-	private String reg_var = "\\$\\{(.*?)\\}";
-	
-	private LinkedList<String> groups = new LinkedList<String>();
+	private SettingLoader settingLoader;
 	
 	public BasicSetting() {
-		this(DEFAULT_CHARSET, false);
 	}
 	
-	/**
-	 * 基本构造<br/>
-	 * 需自定义初始化配置文件<br/>
-	 * 
-	 * @param charset 字符集
-	 * @param isUseVariable 是否使用变量
-	 */
-	public BasicSetting(Charset charset, boolean isUseVariable) {
-		this.charset = charset;
-		this.isUseVariable = isUseVariable;
-	}
-
 	/**
 	 * 构造，使用相对于Class文件根目录的相对路径
 	 * 
@@ -96,6 +65,14 @@ public class BasicSetting extends AbsSetting{
 			throw new RuntimeException(StrUtil.format("Can not find Setting file: [{}]", pathBaseClassLoader));
 		}
 		this.init(url, charset, isUseVariable);
+	}
+	
+	/**
+	 * 构造
+	 * @param pathBaseClassLoader 相对路径（相对于当前项目的classes路径）
+	 */
+	public BasicSetting(String pathBaseClassLoader) {
+		this(pathBaseClassLoader, DEFAULT_CHARSET, false);
 	}
 
 	/**
@@ -146,14 +123,6 @@ public class BasicSetting extends AbsSetting{
 		this.init(url, charset, isUseVariable);
 	}
 	
-	/**
-	 * 构造
-	 * @param pathBaseClassLoader 相对路径（相对于当前项目的classes路径）
-	 */
-	public BasicSetting(String pathBaseClassLoader) {
-		this(pathBaseClassLoader, DEFAULT_CHARSET, false);
-	}
-	
 	/*--------------------------公有方法 start-------------------------------*/
 	/**
 	 * 初始化设定文件
@@ -167,110 +136,21 @@ public class BasicSetting extends AbsSetting{
 		if (settingUrl == null) {
 			throw new RuntimeException("Null setting url or charset define!");
 		}
+		this.settingUrl = settingUrl;
 		this.charset = charset;
 		this.isUseVariable = isUseVariable;
-		this.settingUrl = settingUrl;
 
-		return this.load(settingUrl);
-	}
-
-	/**
-	 * 加载设置文件
-	 * 
-	 * @param settingUrl 配置文件URL
-	 * @return 加载是否成功
-	 */
-	synchronized public boolean load(URL settingUrl) {
-		if (settingUrl == null) {
-			throw new RuntimeException("Null setting url define!");
-		}
-		log.debug("Load setting file [{}]", settingUrl.getPath());
-		InputStream settingStream = null;
-		try {
-			settingStream = settingUrl.openStream();
-			load(settingStream, isUseVariable);
-		} catch (IOException e) {
-			log.error(e, "Load setting error!");
-			return false;
-		} finally {
-			IoUtil.close(settingStream);
-		}
-		return true;
+		return load();
 	}
 
 	/**
 	 * 重新加载配置文件
 	 */
-	public void reload() {
-		this.load(settingUrl);
-	}
-
-	/**
-	 * 加载设置文件。 此方法不会关闭流对象
-	 * 
-	 * @param settingStream 文件流
-	 * @param isUseVariable 是否使用变量（替换配置文件值中含有的变量）
-	 * @return 加载成功与否
-	 * @throws IOException
-	 */
-	public boolean load(InputStream settingStream, boolean isUseVariable) throws IOException {
-		map.clear();
-		BufferedReader reader = null;
-		try {
-			reader = new BufferedReader(new InputStreamReader(settingStream, charset));
-			// 分组
-			String group = null;
-			
-			while (true) {
-				String line = reader.readLine();
-				if (line == null) {
-					break;
-				}
-				line = line.trim();
-				// 跳过注释行和空行
-				if (StrUtil.isBlank(line) || line.startsWith(COMMENT_FLAG_PRE)) {
-					continue;
-				}
-				
-				// 记录分组名
-				if (line.charAt(0) == GROUP_SURROUND[0] && line.charAt(line.length() - 1) == GROUP_SURROUND[1]) {
-					group = line.substring(1, line.length() - 1).trim();
-					this.groups.add(group);
-					continue;
-				}
-				
-				String[] keyValue = line.split(ASSIGN_FLAG, 2);
-				// 跳过不符合简直规范的行
-				if (keyValue.length < 2) {
-					continue;
-				}
-				
-				String key = keyValue[0].trim();
-				if (false == StrUtil.isBlank(group)) {
-					key = group + StrUtil.DOT + key;
-				}
-				String value = keyValue[1].trim();
-				
-				// 替换值中的所有变量变量（变量必须是此行之前定义的变量，否则无法找到）
-				if (isUseVariable) {
-					value = replaceVar(value);
-				}
-				map.put(key, value);
-			}
-		} finally {
-			IoUtil.close(reader);
+	synchronized public boolean load() {
+		if(null == this.settingLoader){
+			settingLoader = new SettingLoader(this, this.charset, this.isUseVariable);
 		}
-		return true;
-	}
-
-	/**
-	 * 设置变量的正则<br/>
-	 * 正则只能有一个group表示变量本身，剩余为字符 例如 \$\{(name)\}表示${name}变量名为name的一个变量表示
-	 * 
-	 * @param regex 正则
-	 */
-	public void setVarRegex(String regex) {
-		this.reg_var = regex;
+		return settingLoader.load(settingUrl);
 	}
 	
 	/**
@@ -375,32 +255,27 @@ public class BasicSetting extends AbsSetting{
 	/**
 	 * 持久化当前设置，会覆盖掉之前的设置<br>
 	 * 持久化会不会保留之前的分组
+	 * @param setting {@link BasicSetting}
 	 * @param absolutePath 设置文件的绝对路径
 	 */
 	public void store(String absolutePath) {
-		Writer writer = null;
-		try {
-			writer = FileUtil.getWriter(absolutePath, charset, false);
-			for (Entry<Object, Object> entry : map.entrySet()) {
-				writer.write(StrUtil.format("{} {} {}", entry.getKey(), ASSIGN_FLAG, entry.getValue()));
-			}
-		} catch (FileNotFoundException e) {
-			throw new RuntimeException(StrUtil.format("Can not find file [{}]!", absolutePath), e);
-		} catch (IOException e) {
-			throw new RuntimeException("Store Setting error!", e);
-		}finally{
-			IoUtil.close(writer);
+		if(null == this.settingLoader){
+			settingLoader = new SettingLoader(this, this.charset, this.isUseVariable);
 		}
+		settingLoader.store(absolutePath);
 	}
-
+	
 	/**
-	 * 存储当前设置，会覆盖掉以前的设置
+	 * 设置变量的正则<br/>
+	 * 正则只能有一个group表示变量本身，剩余为字符 例如 \$\{(name)\}表示${name}变量名为name的一个变量表示
 	 * 
-	 * @param path 相对路径
-	 * @param clazz 相对的类
+	 * @param regex 正则
 	 */
-	public void store(String path, Class<?> clazz) {
-		this.store(FileUtil.getAbsolutePath(path, clazz));
+	public void setVarRegex(String regex) {
+		if(null == this.settingLoader){
+			throw new NullPointerException("SettingLoader is null !");
+		}
+		this.settingLoader.setVarRegex(regex);
 	}
 	
 	/**
@@ -521,25 +396,6 @@ public class BasicSetting extends AbsSetting{
 	}
 
 	/*--------------------------Private Method start-------------------------------*/
-	/**
-	 * 替换给定值中的变量标识
-	 * @param value 值
-	 * @return 替换后的字符串
-	 */
-	private String replaceVar(String value) {
-		// 找到所有变量标识
-		final Set<String> vars = ReUtil.findAll(reg_var, value, 0, new HashSet<String>());
-		for (String var : vars) {
-			// 查找变量名对应的值
-			Object varValue = map.get(ReUtil.get(reg_var, var, 1));
-			if (null != varValue && value instanceof CharSequence) {
-				// 替换标识
-				value = value.replace(var, (CharSequence)varValue);
-			}
-		}
-		return value;
-	}
-	
 	/**
 	 * 组合Key和Group，组合后为group.key
 	 * @param key